requirements.txt
argh
discord.py
mal-api